Bring website up to a deployable beta state
This is a large commit with many changes to the look and operation of the website. An additional Alumni page has been added, and brothers on the Brothers page have been added.

The slide carousle is now fully functioning and adjusts better to mobile screens.

Email and contact functionality has been added by connecting to a php script recycled from previous versions of the OT website.